fixup install process

+++ PKGBUILD
@@ -1,7 +1,7 @@
 # Maintainer: FLX (Paul Werther) <flx@evait.de>
 pkgname=pd-tools
 _pkgname=pdtm
-pkgver=20240124
+pkgver=20240317
 pkgrel=1
 pkgdesc='Provides all open source tools available in one package from ProjectDiscovery by installing the precompiled binaries using pdtm'
 arch=('x86_64' 'armv6h' 'armv7h' 'aarch64')
@@ -41,6 +41,8 @@ package() {
   cd "${srcdir}/${_pkgname}"
   ./cmd/build/${_pkgname} -install-all -binary-path ${pkgdir}
   for file in ${pkgdir}/*; do
-    install -Dvm755 "$file" -t "${pkgdir}/usr/bin/"
+    if [ -f "$file" ]; then
+      install -Dvm755 "$file" -t "${pkgdir}/usr/bin/"
+    fi
   done
 }
